+++ rpc/types/types.go
@@ -187,6 +187,12 @@ type OneFeeHistory struct {
 	GasUsedRatio         float64    // the ratio of gas used to the gas limit for each block
 }
 
+// AccessListResult represents the access list and gas used for a transaction
+type AccessListResult struct {
+	AccessList *ethtypes.AccessList `json:"accessList"`
+	GasUsed    *hexutil.Uint64      `json:"gasUsed"`
+}
+
 type TraceConfig struct {
 	evmtypes.TraceConfig
 	TracerConfig   json.RawMessage `json:"tracerConfig"`